online_exam_api/requirement.txt
cmake==3.22.4
dlib==19.18.0
face-recognition==1.3.0
mediapipe==0.8.9.1
Flask==2.1.1
Flask-RESTful==0.3.9
pandas==1.4.2
pytesseract==0.3.9
blend-modes==2.1.0
Pillow==9.1.0
requests==2.27.1